+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 1.1.0 - 2019-01-03
+
+### Added
+Add the display of the readme for a selected extension.
+
+### Fixed
+Fix refresh extensions on click "Check for update" in activity bar.
+
 ## 1.0.1 - 2019-01-01
 
 ### Fixed